+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Added withFilename method to FilePath and UrlPath classes.
 
 ## [2.1.0] - 2019-06-25
 - Added getPathAndQueryString method to Url class.